fix: treat placeholders as visual element
This changes how we deal with placeholders.

Currently, a placeholder behaves as a visual hint of what we want the
user to type. It also behaves as an optional default value, in that you
can press `<Enter>` or `<Tab>` to insert it as the value.

This is confusing since it means it is a "partial" default value, and a
placeholder at the same time.

This change basically removes the placeholder logic from core, such that
the `text` and `autocomplete` prompts use `placeholder` purely as a
render/visual hint.

This updated behaviour means we render the placeholder when no value is
set, but it can no longer be inserted as a value. To achieve that, you
should combine `defaultValue` with `placeholder` (i.e. set them to the
same value).