fix: do not pass initialValue down
If we pass the `initialValue` of an auto complete prompt to the core, it
will result in setting `this.value` to it.

Normally, that would be ok. However, in the case of auto complete, the
value is an array. All sorts of chaos is caused by this.

For now, we can simply not pass it down. In future, we need to make
`value` strongly typed and no longer mutated until finalisation.