fix(autocomplete): use first option as initial value in non-multiple
When `multiple: false` of an auto complete prompt, we should be setting
the initial value (i.e. the selected value) as the first one in the
list.

This way, when pressing `<RETURN>` without any movement, the result will
be the first value.

When `multiple: true`, this will leave the result as `[]`.

Fixes #331