fix: support off-screen lines when re-rendering
This fixes a deep, difficult bug :eyes:

Let's try explain by example...

If you have an `autocomplete` which has 2 options:

- First option is multiple lines (let's say 6 lines)
- Second option is one line

In a terminal which is 10 rows tall, we are likely to exceed the
terminal height since `6 (row 1) + 1 (row 2) + 4 (message and UI) = 11`.

When this happens, the first render will be fine and the top row will be
off screen.

Now when we navigate (e.g. press `<down>`), the current rendering logic
will assume that the _entire_ frame is on screen. It'll diff the
previous frame and this frame, then update the changed rows.

Because it assumes the entire frame is on screen, it'll try move down to
row `N` (where `N` is the changed row). Here lies the problem: row `N`
will be off by `1` (because `1` row is off screen right now). That means
row `4` of the raw text is visually row `3` in your terminal.

This fix basically accounts for that offset, which fixes a whole bunch
of weird behaviours in smaller terminals, or just long prompts.

One notable thing we also account for in this fix: if the new frame fits
on screen, but the last frame didn't, we erase the entire frame and
render it all to regain screen space.